refactor(export): split GlobalId identification out of merged.rs
The two forked #2951 chains each grew merged.rs independently
(guid-misidentification/schema-coverage added the IfcRoot type check and
the 54-entry legacy rooted-type table; visibility-filter/relationship-
narrowing added MergedOptions.included and narrowing). Combined, the file
crossed the 400-line module-size ratchet (432 lines) with no allowlist row.

Split along the seam merged.rs already documented in its own comments:
GlobalId identification and re-stamping (leading_guid, is_legacy_rooted_type,
replace_leading_guid, deterministic_global_id, mint_unique_guid) is a
self-contained concern independent of the emission loop that calls it
(offsetting, visibility filtering, relationship narrowing) -- the same kind
of seam merged_visibility.rs already split out for the filtering side.

merged.rs: 245 lines. merged_guid.rs: 198 lines. Ratchet passes with no
allowlist changes and no digest change.
